backend/requirements.txt
# Dependencies

# Core Framework
Flask==2.3.3
Werkzeug==2.3.7

# Database
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.20
psycopg2-binary==2.9.7

# Real-time comms
Flask-SocketIO==5.3.4
python-socketio==5.9.0
python-engineio==4.7.1

# CORS Support
Flask-CORS==4.0.0

# Authentication & Security
Flask-JWT-Extended==4.5.2
Flask-Login==0.6.2
PyJWT==2.8.0

# MQTT Protocol
paho-mqtt==1.6.1

# SMS Notifications
twilio==8.10.0

# env. conf.
python-dotenv==1.0.0

# HTTP Requests
requests==2.31.0

# Server & Deployment
gunicorn==21.2.0
eventlet==0.33.3

# Optional: dev & testing
pytest==7.4.0
pytest-flask==1.2.0
black==23.7.0
flake8==6.1.0
